Move delegate table row and header to dedicated files
